dev-1.4.3.24
Events fully implemented.

More info is available in the wiki.

+++ src/main/java/net/mat0u5/lifeseries/seasons/boogeyman/BoogeymanManager.java
@@ -11,6 +11,7 @@ import net.mat0u5.lifeseries.utils.other.TaskScheduler;
 import net.mat0u5.lifeseries.utils.other.TextUtils;
 import net.mat0u5.lifeseries.utils.player.PlayerUtils;
 import net.mat0u5.lifeseries.utils.player.ScoreboardUtils;
+import net.mat0u5.lifeseries.utils.world.DatapackIntegration;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
 import net.minecraft.server.level.ServerPlayer;
@@ -125,6 +126,7 @@ public class BoogeymanManager {
         player.addTag("boogeyman");
         boogeymanChosen = true;
         boogeymanListChanged = true;
+        DatapackIntegration.EVENT_BOOGEYMAN_ADDED.trigger(new DatapackIntegration.Events.MacroEntry("Player", player.getScoreboardName()));
         return newBoogeyman;
     }
 
@@ -141,6 +143,8 @@ public class BoogeymanManager {
         if (boogeyman == null) return;
         boogeymen.remove(boogeyman);
         player.removeTag("boogeyman");
+        player.removeTag("boogeyman_cured");
+        player.removeTag("boogeyman_failed");
         if (boogeymen.isEmpty()) boogeymanChosen = false;
         player.sendSystemMessage(Component.nullToEmpty("§c [NOTICE] You are no longer a Boogeyman!"));
     }
@@ -166,6 +170,8 @@ public class BoogeymanManager {
         }
         boogeyman.failed = false;
         boogeyman.cured = false;
+        player.removeTag("boogeyman_cured");
+        player.removeTag("boogeyman_failed");
         boogeyman.died = false;
         boogeyman.resetKills();
     }
@@ -175,6 +181,8 @@ public class BoogeymanManager {
         Boogeyman boogeyman = getBoogeyman(player);
         if (boogeymen == null) return;
         boogeyman.failed = false;
+        player.addTag("boogeyman_cured");
+        player.removeTag("boogeyman_failed");
         if (boogeyman.cured) return;
         boogeyman.cured = true;
         PlayerUtils.sendTitle(player,Component.nullToEmpty("§aYou are cured!"), 20, 30, 20);
@@ -190,7 +198,8 @@ public class BoogeymanManager {
                 PlayerUtils.broadcastMessage(TextUtils.format("{}§7 is cured of the Boogeyman curse!", player));
             }
         }
-        if (stealLife) {
+        DatapackIntegration.EVENT_BOOGEYMAN_CURE_REWARD.trigger(new DatapackIntegration.Events.MacroEntry("Player", player.getScoreboardName()));
+        if (!DatapackIntegration.EVENT_BOOGEYMAN_CURE_REWARD.isCanceled() && stealLife) {
             player.ls$addLife();
         }
     }
@@ -402,29 +411,35 @@ public class BoogeymanManager {
         Boogeyman boogeyman = getBoogeyman(player);
         if (boogeymen == null) return false;
 
+        player.removeTag("boogeyman_cured");
+        player.addTag("boogeyman_failed");
         boogeyman.cured = false;
         if (boogeyman.failed) return false;
         boogeyman.failed = true;
 
         boolean canChangeLives = player.ls$isAlive() && !player.ls$isOnLastLife(true);
 
-        if (BOOGEYMAN_ADVANCED_DEATHS) {
-            PlayerUtils.sendTitle(player,Component.nullToEmpty("§cThe curse consumes you.."), 20, 30, 20);
-            if (BOOGEYMAN_ANNOUNCE_OUTCOME && sendMessage) {
-                PlayerUtils.broadcastMessage(TextUtils.format("{}§7 failed to kill a player while being the §cBoogeyman§7. They have been consumed by the curse.", player));
+
+        DatapackIntegration.EVENT_BOOGEYMAN_FAIL_REWARD.trigger(new DatapackIntegration.Events.MacroEntry("Player", player.getScoreboardName()));
+        if (!DatapackIntegration.EVENT_BOOGEYMAN_FAIL_REWARD.isCanceled()) {
+            if (BOOGEYMAN_ADVANCED_DEATHS) {
+                PlayerUtils.sendTitle(player,Component.nullToEmpty("§cThe curse consumes you.."), 20, 30, 20);
+                if (BOOGEYMAN_ANNOUNCE_OUTCOME && sendMessage) {
+                    PlayerUtils.broadcastMessage(TextUtils.format("{}§7 failed to kill a player while being the §cBoogeyman§7. They have been consumed by the curse.", player));
+                }
+                if (canChangeLives) {
+                    AdvancedDeathsManager.setPlayerLives(player, 1);
+                }
             }
-            if (canChangeLives) {
-                AdvancedDeathsManager.setPlayerLives(player, 1);
-            }
-        }
-        else {
-            PlayerUtils.sendTitle(player,Component.nullToEmpty("§cYou have failed."), 20, 30, 20);
-            PlayerUtils.playSoundToPlayer(player, SoundEvent.createVariableRangeEvent(IdentifierHelper.vanilla("lastlife_boogeyman_fail")));
-            if (BOOGEYMAN_ANNOUNCE_OUTCOME && sendMessage) {
-                PlayerUtils.broadcastMessage(TextUtils.format("{}§7 failed to kill a player while being the §cBoogeyman§7. They have been dropped to their §cLast Life§7.", player));
-            }
-            if (canChangeLives) {
-                player.ls$setLives(1);
+            else {
+                PlayerUtils.sendTitle(player,Component.nullToEmpty("§cYou have failed."), 20, 30, 20);
+                PlayerUtils.playSoundToPlayer(player, SoundEvent.createVariableRangeEvent(IdentifierHelper.vanilla("lastlife_boogeyman_fail")));
+                if (BOOGEYMAN_ANNOUNCE_OUTCOME && sendMessage) {
+                    PlayerUtils.broadcastMessage(TextUtils.format("{}§7 failed to kill a player while being the §cBoogeyman§7. They have been dropped to their §cLast Life§7.", player));
+                }
+                if (canChangeLives) {
+                    player.ls$setLives(1);
+                }
             }
         }
         return true;
@@ -532,6 +547,8 @@ public class BoogeymanManager {
         ServerPlayer player = boogeyman.getPlayer();
         if (player != null) {
             player.removeTag("boogeyman");
+            player.removeTag("boogeyman_cured");
+            player.removeTag("boogeyman_failed");
         }
         TaskScheduler.scheduleTask(100, this::chooseNewBoogeyman);
     }
@@ -562,6 +579,7 @@ public class BoogeymanManager {
             ServerPlayer player = boogeyman.getPlayer();
             if (player != null) {
                 if (!playerFailBoogeyman(player, true)) {
+                    player.addTag("boogeyman_failed");
                     boogeyman.failed = true;
                 }
             }
